updated plot4.R

+++ plot4.R
@@ -1,15 +1,20 @@
-## Script to create plot4.png
-## This was created as part of Coursera Data Science Specialization
-## Course: Exploratory Data Analysis, Course Project # 1
-## Author: Brent Brewington
+## Coursera - Exploratory Data Analysis
+## Course Project # 1
+## Author: Brent Brewington, (github: bbrewington)
+## Plot4.R
 
 # Get data from file "household_power_consumption", and save to data frame "DF"
 
-temp <- tempfile()
-download.file("https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2Fhousehold_power_consumption.zip",temp)
-DF <- read.table(unz(temp, "household_power_consumption.txt"), 
-                 header=TRUE,sep=";",na.strings="?",stringsAsFactors=FALSE)
-unlink(temp)
+if(!("household_power_consumption.txt" %in% list.files())){
+  temp <- tempfile()
+  download.file("https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2Fhousehold_power_consumption.zip",temp)
+  DF <- read.table(unz(temp, "household_power_consumption.txt"), 
+                   header=TRUE,sep=";",na.strings="?",stringsAsFactors=FALSE)
+  unlink(temp)
+} else{
+  DF <- read.table("household_power_consumption.txt", 
+                   header=TRUE,sep=";",na.strings="?",stringsAsFactors=FALSE)
+}
 
 # Create new data frame "DF_subset", which only includes Feb 1, 2007 - Feb 2, 2007
 
@@ -30,13 +35,14 @@ par(mfrow=c(2,2))
   #Add plot to row 1, col 2
   plot(DateTime,DF_subset$Voltage,xlab="datetime",ylab="Voltage", yaxp=c(234,246,3),type="n")
   lines(DateTime,DF_subset$Voltage)
+  axis(side=2, at = seq(234,246,2))
   #Add plot to row 2, col 1
   with(DF_subset, plot(DateTime, Sub_metering_1, ylab="Energy sub metering", xlab="",type = "n"))
   with(DF_subset, lines(DateTime, Sub_metering_1, col = "black"))
   with(DF_subset, lines(DateTime, Sub_metering_2, col = "red"))
   with(DF_subset, lines(DateTime, Sub_metering_3, col = "blue"))
   legend("topright",legend = c("Sub_metering_1","Sub_metering_2","Sub_metering_3"), 
-         lty=c(1,1,1), col=c("black","red","blue"), cex=.7)
+         lty=c(1,1,1), col=c("black","red","blue"), cex=.9, bty="n")
   #Add plot to row 2, col 2
   plot(DateTime,DF_subset$Global_reactive_power,xlab="datetime",ylab="Global_reactive_power", yaxp=c(0.0,0.5,5),type="n")
   lines(DateTime,DF_subset$Global_reactive_power)
